docker compose setup complete

+++ src/app.py
@@ -8,29 +8,29 @@ app = Flask(__name__)
 
    
 
-mysql_config = {
+
+def create_db_connection():
+    mysql_config = {
     'host': os.environ.get('MYSQL_HOST'),
     'user': os.environ.get('MYSQL_USER'),
     'port': os.environ.get('MYSQL_PORT'),
     'password':  os.environ.get('MYSQL_PASSWORD'),
     'database':  os.environ.get('MYSQL_DB'),
-}
-cursor = None
-try:
-    conn = mysql.connector.connect(**mysql_config)
+    }
 
-    print("MySQL connection successful!")
-    cursor = conn.cursor()
+    try:
+        conn = mysql.connector.connect(**mysql_config)
 
+        print("MySQL connection successful!")
+        return conn 
 
-    
+    except mysql.connector.Error as err:
 
-except mysql.connector.Error as err:
-
-    print(f"MySQL connection error: {err}")
+        print(f"MySQL connection error: {err}")
 
 @app.route('/customers', methods=['POST'])
 def create_customer():
+    
     data = request.json
     if all(key not in data for key in ['name', 'phone_number', 'email']):
         return jsonify({"error": "Name, phone number, and email are required"}), 400
@@ -38,36 +38,51 @@ def create_customer():
     name = data.get('name')
     phone_number = data.get('phone_number')
     email = data.get('email')
-    try:
-
-        cursor.execute("INSERT INTO CUSTOMER (name, phone_number, email) VALUES (%s, %s, %s)", (name, phone_number, email))
-        lastrowid = cursor.lastrowid
-        conn.commit()
-
-    except mysql.connector.Error as e:
-        conn.rollback()
-        return jsonify({"error": e.msg})
+    
+    conn = create_db_connection()
+    
+    if conn is not None:
         
-    return jsonify({"message": f"Customer added successfully with values {name, phone_number, email} with ID: {lastrowid}"}), 201
+        try:
+            cursor = conn.cursor()
+            cursor.execute("INSERT INTO CUSTOMER (name, phone_number, email) VALUES (%s, %s, %s)", (name, phone_number, email))
+            lastrowid = cursor.lastrowid
+            conn.commit()
+
+            cursor.close()
+            conn.close()     
+        except mysql.connector.Error as e:
+            conn.rollback()
+            return jsonify({"error": e.msg})
+            
+        return jsonify({"message": f"Customer added successfully with values {name, phone_number, email} with ID: {lastrowid}"}), 201
+    return jsonify({"error": "An unexpected error occurred"}), 500
 
 @app.route('/customers/<int:customer_id>', methods=['GET'])
 def get_customer(customer_id):
 
-    try: 
-        cursor.execute("SELECT * FROM CUSTOMER WHERE customer_id = %s", (customer_id,))
-        customer = cursor.fetchone()
-        print(customer)
-        if customer:
-            customer_data = {
-                "id": customer[0],
-                "name": customer[1],
-                "phone_number": customer[2],
-                "email": customer[3]
-            }
-            return jsonify(customer_data)
-    except Exception as e:
-        return jsonify({"error": f"Failed to retrieve customer with id {customer_id}"}), 404
-    return jsonify({"error": "Customer not found"}), 404
+    conn = create_db_connection()
+    
+    if conn is not None:
+        try:
+            cursor = conn.cursor() 
+            cursor.execute("SELECT * FROM CUSTOMER WHERE customer_id = %s", (customer_id,))
+            customer = cursor.fetchone()
+            cursor.close()
+            conn.close()     
+            if customer:
+                customer_data = {
+                    "id": customer[0],
+                    "name": customer[1],
+                    "phone_number": customer[2],
+                    "email": customer[3]
+                }
+                return jsonify(customer_data)
+        except Exception as e:
+            print(e.msg)
+            return jsonify({"error": f"Failed to retrieve customer with id {customer_id}"}), 404
+
+    return jsonify({"error": "An unexpected error occurred"}), 500
 
 @app.route('/customers/<int:customer_id>', methods=['PUT'])
 def update_customer(customer_id):
@@ -96,25 +111,40 @@ def update_customer(customer_id):
     # remove trailing whitespace and comma
     query = query.rstrip().rstrip(',') 
     
-    query += " WHERE id = %s"
+    query += " WHERE customer_id = %s"
     values.append(customer_id)
 
 
-    try:
-        cursor.execute(query, tuple(values))
-        conn.commit()
-        return jsonify({"message": f"Customer updated successfully with {tuple(values)}"}), 201
-    except Exception as e:
-        return jsonify({"error": f"Failed to update customer with {tuple(values)}"}), 422
+    conn = create_db_connection()
+    if conn is not None:
+        try:
+            cursor = conn.cursor()
+            cursor.execute(query, tuple(values))
+            conn.commit()
+            cursor.close()
+            conn.close()  
+            return jsonify({"message": f"Customer updated successfully with {tuple(values)}"}), 201
+        except Exception as e:
+            print(e.msg)
+            return jsonify({"error": f"Failed to update customer with {tuple(values)}"}), 422
+    return jsonify({"error": "An unexpected error occurred"}), 500
 
 @app.route('/customers/<int:customer_id>', methods=['DELETE'])
 def delete_customer(customer_id):
-    try:   
-        cursor.execute("DELETE FROM CUSTOMER WHERE customer_id = %s", (customer_id,))
-        conn.commit()
-        return jsonify({"message": f"Customer with id {customer_id} deleted successfully"})
-    except Exception as e:
-        return jsonify({"error": f"Failed to delete customer with id {customer_id}"}), 422
+    
+    conn = create_db_connection()
+    
+    if conn is not None:
+        try:
+            cursor = conn.cursor()
+            cursor.execute("DELETE FROM CUSTOMER WHERE customer_id = %s", (customer_id,))
+            conn.commit()
+            cursor.close()
+            conn.close()  
+            return jsonify({"message": f"Customer with id {customer_id} deleted successfully"})
+        except Exception as e:
+            return jsonify({"error": f"Failed to delete customer with id {customer_id}"}), 422
+    return jsonify({"error": "An unexpected error occurred"}), 500
 
 if __name__ == '__main__':
     app.run(host='0.0.0.0',debug=True, port=int(os.environ.get('APP_PORT')))
